gemma/requirements.txt
# Core ML / Transformers
torch # torchvision/torchaudio requires this! Version selected by pip might vary.
torchvision
torchaudio
transformers
bitsandbytes
peft
accelerate

# Image & Numerics
pillow
numpy

# API / Server
fastapi
uvicorn

# Optional / Performance (CUDA & specific GPU needed)
flash-attn